ai-agents/pyproject.toml
[project]
name = "watch2give-ai-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "langchain>=0.3.23",
    "langchain-groq>=0.3.2",
    "langgraph>=0.3.30",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.1",
]
